Indy 894 Update verbiage in Getting Started Guide (#391)
Updates verbiage in Getting Started Guide to reflect small changes due
to removing branding from Sovrin to Indy. I also updated verbiage
throughout to reflect actors specifically, where it made sense to do
so for the sake of clarity. Example, instead of just saying "Indy", I
changed it to "Indy code base" or "the ledger" or "the network."

Signed-off-by: TechWritingWhiz <TechWritingWhiz@gmail.com>